backend3/requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
websockets==11.0.3
python-multipart==0.0.6
motor==3.3.1
pymongo==4.5.0
pyjwt==2.8.0
bcrypt==4.0.1
passlib==1.7.4
python-dotenv==1.0.0
pydantic==2.4.2
email-validator==2.0.0.post2